+++ CHANGELOG.rst
@@ -1,5 +1,11 @@
 Changelog for Mantarray Frontend Components
 ===========================================
+0.7.6 (2022-06-24)
+
+Added:
+
+- Added a test file for the new h5 warning in pages
+- added front-end h5 warning
 
 0.7.5 (2022-06-16)
 ------------------